+++ src/test/java/uk/gov/pay/ledger/pact/event/CancelledByUserEventQueueContractTest.java
@@ -1,13 +1,15 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
+import com.google.gson.Gson;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
+import uk.gov.pay.ledger.event.model.SalientEventType;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
@@ -22,13 +24,13 @@ import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
-import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class Gateway3dsExemptionResultObtainedEventQueueContractTest {
+public class CancelledByUserEventQueueContractTest {
+    Gson gson = new Gson();
+
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -38,45 +40,46 @@ public class Gateway3dsExemptionResultObtainedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "exemptionResult_externalId";
+    private String externalId = "externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
+    private String gatewayAccountId = "gateway_account_id";
+    private String gatewayTransactionId = "validGatewayTransactionId";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createGateway3dsExemptionResultObtainedEventPact(MessagePactBuilder builder) {
-        String gateway3dsExemptionResultObtainedEvent = "GATEWAY_3DS_EXEMPTION_RESULT_OBTAINED";
-        QueuePaymentEventFixture gateway3dsExemptionResultObtained = aQueuePaymentEventFixture()
+    public MessagePact createCancelledByUserEventPact(MessagePactBuilder builder) {
+        QueuePaymentEventFixture canceledByUserEventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withEventType(gateway3dsExemptionResultObtainedEvent)
-                .withDefaultEventDataForEventType(gateway3dsExemptionResultObtainedEvent)
+                .withGatewayAccountId(gatewayAccountId)
+                .withEventType(SalientEventType.CANCELLED_BY_USER.toString())
+                .withDefaultEventDataForEventType(SalientEventType.CANCELLED_BY_USER.toString())
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a gateway 3DS exemption result obtained message")
-                .withMetadata(metadata)
-                .withContent(gateway3dsExemptionResultObtained.getAsPact())
+                .expectsToReceive("a cancelled by user message")
+                .withContent(canceledByUserEventFixture.getAsPact())
                 .toPact();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
-        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
+
+        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-                && transactionDao.findTransactionByExternalId(externalId).get().getTransactionDetails().contains("\"exemption3ds\""));
+        );
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
-
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"exemption3ds\": \"HONOURED\""));
+        Map<String, String> transactionDetails = gson.fromJson(transaction.get().getTransactionDetails(), Map.class);
+        assertThat(transactionDetails.get("gateway_transaction_id"), is(gatewayTransactionId));
     }
 
     public void setMessage(byte[] messageContents) {
