+++ src/test/java/uk/gov/pay/ledger/pact/event/FeeIncurredEventQueueContractTest.java
@@ -1,36 +1,38 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
-import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
-import uk.gov.pay.ledger.event.model.SalientEventType;
+import uk.gov.pay.ledger.event.dao.EventDao;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
+import uk.gov.pay.ledger.util.DatabaseTestHelper;
 import uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture;
 
 import java.time.ZonedDateTime;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class CaptureSubmittedEventQueueContractTest {
-    Gson gson = new Gson();
-
+public class FeeIncurredEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -39,29 +41,47 @@ public class CaptureSubmittedEventQueueContractTest {
             config("queueMessageReceiverConfig.backgroundProcessingEnabled", "true")
     );
 
+    private GsonBuilder gsonBuilder = new GsonBuilder();
     private byte[] currentMessage;
-    private String externalId = "externalId";
+    private String paymentExternalId = "payment-external-id";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
-    private String gatewayAccountId = "gateway_account_id";
+
+    public static final long fee = 33L;
+    public static final long netAmount = 67L;
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentCreatedEventPact(MessagePactBuilder builder) {
-        QueuePaymentEventFixture paymentCreatedEventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
-                .withResourceExternalId(externalId)
+    public MessagePact createFeeIncurredEventPact(MessagePactBuilder builder) {
+        String eventData = gsonBuilder.create()
+                .toJson(Map.of(
+                        "fee", fee,
+                        "net_amount", netAmount,
+                        "fee_breakdown", List.of(
+                                Map.of(
+                                        "amount", 24,
+                                        "fee_type", "transaction"
+                                )
+                        )
+                ));
+        QueuePaymentEventFixture eventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
+                .withResourceExternalId(paymentExternalId)
                 .withEventDate(eventDate)
-                .withGatewayAccountId(gatewayAccountId)
-                .withEventType(SalientEventType.CAPTURE_SUBMITTED.toString())
-                .withDefaultEventDataForEventType(SalientEventType.CAPTURE_SUBMITTED.toString())
-                .withLive(true);
+                .withServiceId(null)
+                .withEventData(eventData)
+                .withEventType("FEE_INCURRED");
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a capture submitted message")
+                .expectsToReceive("a fee incurred message")
                 .withMetadata(metadata)
-                .withContent(paymentCreatedEventFixture.getAsPact())
+                .withContent(eventFixture.getAsPact())
                 .toPact();
+    }
+
+    @Before
+    public void setUp() {
+        DatabaseTestHelper.aDatabaseTestHelper(appRule.getJdbi()).truncateAllData();
     }
 
     @Test
@@ -72,14 +92,14 @@ public class CaptureSubmittedEventQueueContractTest {
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
 
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+                () -> transactionDao.findTransactionByExternalId(paymentExternalId).isPresent()
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(paymentExternalId);
         assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getExternalId(), is(externalId));
-        Map<String, String> transactionDetails = gson.fromJson(transaction.get().getTransactionDetails(), Map.class);
-        assertThat(transactionDetails.get("capture_submitted_date"), is(eventDate.toString()));
+        assertThat(transaction.get().getFee(), is(fee));
+        assertThat(transaction.get().getNetAmount(), is(netAmount));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"fee_breakdown\": [{\"amount\": 24, \"fee_type\": \"transaction\"}]}"));
     }
 
     public void setMessage(byte[] messageContents) {
