+++ src/test/java/uk/gov/pay/ledger/pact/event/PaymentCreatedEventQueueContractTest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
@@ -8,6 +8,7 @@ import au.com.dius.pact.model.v3.messaging.MessagePact;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
+import uk.gov.pay.ledger.event.dao.EventDao;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
@@ -26,9 +27,8 @@ import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
-import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class PaymentDetailsEnteredEventQueueContractTest {
+public class PaymentCreatedEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -38,69 +38,62 @@ public class PaymentDetailsEnteredEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "paymentDetails_externalId";
+    private String externalId = "created_externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
+    private String gatewayAccountId = "gateway_account_id";
+    private String credentialExternalId = "credential-external-id-1";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentDetailsEnteredEventPact(MessagePactBuilder builder) {
-        String paymentDetailsEnteredEventName = "PAYMENT_DETAILS_ENTERED";
-        QueuePaymentEventFixture paymentDetailsEntered = aQueuePaymentEventFixture()
+    public MessagePact createPaymentCreatedEventPact(MessagePactBuilder builder) {
+        QueuePaymentEventFixture paymentCreatedEventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withEventType(paymentDetailsEnteredEventName)
-                .withGatewayAccountId("I0YI1")
-                .withDefaultEventDataForEventType(paymentDetailsEnteredEventName)
+                .withGatewayAccountId(gatewayAccountId)
+                .withCredentialExternalId(credentialExternalId)
+                .withDefaultEventDataForEventType("PAYMENT_CREATED")
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payment details entered message")
+                .expectsToReceive("a payment created message")
                 .withMetadata(metadata)
-                .withContent(paymentDetailsEntered.getAsPact())
+                .withContent(paymentCreatedEventFixture.getAsPact())
                 .toPact();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
-        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        setupTransaction(transactionDao);
-
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
+        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
+        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
+
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-                        && transactionDao.findTransactionByExternalId(externalId).get().getCardholderName() != null
+                () -> transactionDao.findTransactionByExternalIdAndGatewayAccountId(externalId, gatewayAccountId).isPresent()
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
-
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalIdAndGatewayAccountId(externalId, gatewayAccountId);
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getCardBrand(), is("visa"));
-        assertThat(transaction.get().getFirstDigitsCardNumber(), is("424242"));
-        assertThat(transaction.get().getLastDigitsCardNumber(), is("4242"));
+        assertThat(transaction.get().getAmount(), is(1000L));
+        assertThat(transaction.get().getDescription(), is("a description"));
+        assertThat(transaction.get().getReference(), is("aref"));
+        assertThat(transaction.get().getGatewayAccountId(), is(gatewayAccountId));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"return_url\": \"https://example.org\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"payment_provider\": \"sandbox\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"language\": \"en\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"delayed_capture\": false"));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"external_metadata\": {\"key\": \"value\"}"));
+        assertThat(transaction.get().getEmail(), is("j.doe@example.org"));
         assertThat(transaction.get().getCardholderName(), is("J citizen"));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"corporate_surcharge\": 5"));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"expiry_date\": \"11/21\""));
         assertThat(transaction.get().getTransactionDetails(), containsString("\"address_line1\": \"12 Rouge Avenue\""));
         assertThat(transaction.get().getTransactionDetails(), containsString("\"address_postcode\": \"N1 3QU\""));
         assertThat(transaction.get().getTransactionDetails(), containsString("\"address_city\": \"London\""));
         assertThat(transaction.get().getTransactionDetails(), containsString("\"address_country\": \"GB\""));
-        assertThat(transaction.get().getTotalAmount(), is(1005L));
-    }
-
-    private void setupTransaction(TransactionDao transactionDao) {
-        aQueuePaymentEventFixture()
-                .withEventType("PAYMENT_CREATED")
-                .withResourceExternalId(externalId)
-                .withEventDate(eventDate.minusHours(1))
-                .insert(appRule.getSqsClient());
-        await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-        );
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"credential_external_id\": \"" + credentialExternalId + "\""));
     }
 
     public void setMessage(byte[] messageContents) {
