+++ src/test/java/uk/gov/pay/ledger/pact/event/PaymentIncludedInPayoutEventQueueContractTest.java
@@ -1,21 +1,21 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
-import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
-import com.google.gson.Gson;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
 import uk.gov.pay.ledger.event.dao.EventDao;
-import uk.gov.pay.ledger.event.model.Event;
+import uk.gov.pay.ledger.event.model.ResourceType;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
+import uk.gov.pay.ledger.util.DatabaseTestHelper;
 import uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture;
 
 import java.time.ZonedDateTime;
@@ -30,8 +30,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-
-public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueContractTest {
+public class PaymentIncludedInPayoutEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -41,30 +40,32 @@ public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueContractTest
     );
 
     private byte[] currentMessage;
-    private String externalId = "captureConfirm_externalId";
+    private String paymentExternalId = "payment-external-id";
+    private String payoutId = "po_12345";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
-    private String gatewayAccountId = "gateway_account_id";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createCaptureConfirmedEventPact(MessagePactBuilder builder) {
-        String eventType = "STATUS_CORRECTED_TO_CAPTURED_TO_MATCH_GATEWAY_STATUS";
-        QueuePaymentEventFixture statusCorrectedToCapturedEvent = QueuePaymentEventFixture.aQueuePaymentEventFixture()
-                .withResourceExternalId(externalId)
+    public MessagePact createPaymentIncludedInPayoutEventPact(MessagePactBuilder builder) {
+        QueuePaymentEventFixture eventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
+                .withResourceExternalId(paymentExternalId)
                 .withEventDate(eventDate)
-                .withGatewayAccountId(gatewayAccountId)
-                .withEventType(eventType)
-                .withDefaultEventDataForEventType(eventType);
+                .withServiceId(null)
+                .withEventData(String.format("{\"gateway_payout_id\": \"%s\"}", payoutId))
+                .withEventType("PAYMENT_INCLUDED_IN_PAYOUT");
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
-        PactDslJsonBody pactBody = statusCorrectedToCapturedEvent.getAsPact();
-
         return builder
-                .expectsToReceive("a status corrected to captured event")
+                .expectsToReceive("a payment included in payout message")
                 .withMetadata(metadata)
-                .withContent(pactBody)
+                .withContent(eventFixture.getAsPact())
                 .toPact();
+    }
+
+    @Before
+    public void setUp() {
+        DatabaseTestHelper.aDatabaseTestHelper(appRule.getJdbi()).truncateAllData();
     }
 
     @Test
@@ -76,22 +77,12 @@ public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueContractTest
         EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
 
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+                () -> transactionDao.findTransactionByExternalId(paymentExternalId).isPresent()
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(paymentExternalId);
         assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getFee(), is(5L));
-        assertThat(transaction.get().getNetAmount(), is(1069L));
-
-        Map<String, Object> transactionDetails = new Gson().fromJson(transaction.get().getTransactionDetails(), Map.class);
-        assertThat(transactionDetails.get("captured_date"), is("2018-03-12T16:25:01.123456Z"));
-
-        Event event = eventDao.getEventsByResourceExternalId(externalId).get(0);
-        Map<String, Object> eventDetails = new Gson().fromJson(event.getEventData(), Map.class);
-        assertThat(eventDetails.get("gateway_event_date"),  is("2018-03-12T16:25:01.123456Z"));
-        assertThat(eventDetails.get("captured_date"),  is("2018-03-12T16:25:01.123456Z"));
+        assertThat(transaction.get().getGatewayPayoutId(), is(payoutId));
     }
 
     public void setMessage(byte[] messageContents) {
