+++ src/test/java/uk/gov/pay/ledger/pact/event/PaymentNotificationCreatedEventQueueContractTest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
@@ -28,7 +28,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class PaymentCreatedEventQueueContractTest {
+public class PaymentNotificationCreatedEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -38,28 +38,30 @@ public class PaymentCreatedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "created_externalId";
+    private String externalId = "notifications_externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
     private String gatewayAccountId = "gateway_account_id";
     private String credentialExternalId = "credential-external-id-1";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentCreatedEventPact(MessagePactBuilder builder) {
-        QueuePaymentEventFixture paymentCreatedEventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
+    public MessagePact createPaymentNotificationCreatedEventPact(MessagePactBuilder builder) {
+        String paymentDetailsEnteredEventName = "PAYMENT_NOTIFICATION_CREATED";
+        QueuePaymentEventFixture paymentNotificationCreatedEventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
                 .withGatewayAccountId(gatewayAccountId)
                 .withCredentialExternalId(credentialExternalId)
-                .withDefaultEventDataForEventType("PAYMENT_CREATED")
+                .withEventType(paymentDetailsEnteredEventName)
+                .withDefaultEventDataForEventType(paymentDetailsEnteredEventName)
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payment created message")
+                .expectsToReceive("a payment notification created message")
                 .withMetadata(metadata)
-                .withContent(paymentCreatedEventFixture.getAsPact())
+                .withContent(paymentNotificationCreatedEventFixture.getAsPact())
                 .toPact();
     }
 
@@ -70,29 +72,34 @@ public class PaymentCreatedEventQueueContractTest {
 
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
         EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
-
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalIdAndGatewayAccountId(externalId, gatewayAccountId).isPresent()
+                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalIdAndGatewayAccountId(externalId, gatewayAccountId);
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getAmount(), is(1000L));
-        assertThat(transaction.get().getDescription(), is("a description"));
-        assertThat(transaction.get().getReference(), is("aref"));
-        assertThat(transaction.get().getGatewayAccountId(), is(gatewayAccountId));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"return_url\": \"https://example.org\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"payment_provider\": \"sandbox\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"language\": \"en\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"delayed_capture\": false"));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"external_metadata\": {\"key\": \"value\"}"));
         assertThat(transaction.get().getEmail(), is("j.doe@example.org"));
+        assertThat(transaction.get().getCardBrand(), is("visa"));
+        assertThat(transaction.get().getFirstDigitsCardNumber(), is("424242"));
+        assertThat(transaction.get().getLastDigitsCardNumber(), is("4242"));
         assertThat(transaction.get().getCardholderName(), is("J citizen"));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"address_line1\": \"12 Rouge Avenue\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"address_postcode\": \"N1 3QU\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"address_city\": \"London\""));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"address_country\": \"GB\""));
+        assertThat(transaction.get().getReference(), is("MRPC12345"));
+        assertThat(transaction.get().getCardBrand(), is("visa"));
+        assertThat(transaction.get().getDescription(), is("New passport application"));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"gateway_transaction_id\": \"providerId\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"amount\": 1000"));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"expiry_date\": \"11/21\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"status\": \"success\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"auth_code\": \"authCode\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"processor_id\": \"processorId\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"telephone_number\": \"+447700900796\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"authorised_date\": \"2018-02-21T16:05:33Z\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"created_date\": \"2018-02-21T15:05:13Z\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"card_brand_label\": \"Visa\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"payment_provider\": \"sandbox\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"expiry_date\": \"11/21\""));
         assertThat(transaction.get().getTransactionDetails(), containsString("\"credential_external_id\": \"" + credentialExternalId + "\""));
     }
