+++ src/test/java/uk/gov/pay/ledger/pact/event/PayoutPaidEventQueueContractTest.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
@@ -30,7 +30,7 @@ import static uk.gov.pay.ledger.payout.state.PayoutState.PAID_OUT;
 import static uk.gov.pay.ledger.payout.state.PayoutState.UNDEFINED;
 import static uk.gov.pay.ledger.util.fixture.PayoutFixture.PayoutFixtureBuilder.aPayoutFixture;
 
-public class PayoutUpdatedEventQueueContractTest {
+public class PayoutPaidEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -40,12 +40,12 @@ public class PayoutUpdatedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String gatewayPayoutId = "po_updated_1234567890";
+    private String gatewayPayoutId = "po_paid_1234567890";
     private ZonedDateTime eventDate = parse("2020-05-13T18:50:00Z");
-    private String eventType = "PAYOUT_UPDATED";
+    private String eventType = "PAYOUT_PAID";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPayoutUpdatedEventPact(MessagePactBuilder builder) {
+    public MessagePact createPayoutPaidEventPact(MessagePactBuilder builder) {
         QueuePayoutEventFixture payoutEventFixture = QueuePayoutEventFixture.aQueuePayoutEventFixture()
                 .withResourceExternalId(gatewayPayoutId)
                 .withEventDate(eventDate)
@@ -56,7 +56,7 @@ public class PayoutUpdatedEventQueueContractTest {
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payout updated message")
+                .expectsToReceive("a payout paid message")
                 .withMetadata(metadata)
                 .withContent(payoutEventFixture.getAsPact())
                 .toPact();
@@ -75,10 +75,11 @@ public class PayoutUpdatedEventQueueContractTest {
         Optional<PayoutEntity> payoutEntity = payoutDao.findByGatewayPayoutId(gatewayPayoutId);
         assertThat(payoutEntity.isPresent(), is(true));
         assertThat(payoutEntity.get().getGatewayPayoutId(), is(gatewayPayoutId));
-        assertThat(payoutEntity.get().getState(), is(UNDEFINED));
+        assertThat(payoutEntity.get().getPaidOutDate().toString(), is("2020-05-13T18:50Z"));
+        assertThat(payoutEntity.get().getState(), is(PAID_OUT));
 
         Map<String, Object> payoutDetails = new Gson().fromJson(payoutEntity.get().getPayoutDetails(), Map.class);
-        assertThat(payoutDetails.get("gateway_status"), is("pending"));
+        assertThat(payoutDetails.get("gateway_status"), is("paid"));
     }
 
     public void setMessage(byte[] messageContents) {
