+++ src/test/java/uk/gov/pay/ledger/pact/event/UserEmailCollectedEventQueueContractTest.java
@@ -1,21 +1,17 @@
-package uk.gov.pay.ledger.pact;
+package uk.gov.pay.ledger.pact.event;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
-import uk.gov.pay.ledger.event.dao.EventDao;
-import uk.gov.pay.ledger.event.model.ResourceType;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
-import uk.gov.pay.ledger.util.DatabaseTestHelper;
 import uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture;
 
 import java.time.ZonedDateTime;
@@ -29,8 +25,9 @@ import static org.awaitility.Awaitility.await;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
+import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class PaymentIncludedInPayoutEventQueueContractTest {
+public class UserEmailCollectedEventQueueContractTest {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -40,49 +37,46 @@ public class PaymentIncludedInPayoutEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String paymentExternalId = "payment-external-id";
-    private String payoutId = "po_12345";
+    private String externalId = "userEmailCol_externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentIncludedInPayoutEventPact(MessagePactBuilder builder) {
-        QueuePaymentEventFixture eventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
-                .withResourceExternalId(paymentExternalId)
+    public MessagePact createUserEmailCollectedEventPact(MessagePactBuilder builder) {
+        String userEmailCollectedEvent = "USER_EMAIL_COLLECTED";
+        QueuePaymentEventFixture paymentDetailsEntered = aQueuePaymentEventFixture()
+                .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withServiceId(null)
-                .withEventData(String.format("{\"gateway_payout_id\": \"%s\"}", payoutId))
-                .withEventType("PAYMENT_INCLUDED_IN_PAYOUT");
+                .withEventType(userEmailCollectedEvent)
+                .withDefaultEventDataForEventType(userEmailCollectedEvent)
+                .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payment included in payout message")
+                .expectsToReceive("a user email collected message")
                 .withMetadata(metadata)
-                .withContent(eventFixture.getAsPact())
+                .withContent(paymentDetailsEntered.getAsPact())
                 .toPact();
-    }
-
-    @Before
-    public void setUp() {
-        DatabaseTestHelper.aDatabaseTestHelper(appRule.getJdbi()).truncateAllData();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
+        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
+
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
-        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
-
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(paymentExternalId).isPresent()
+                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+                        && transactionDao.findTransactionByExternalId(externalId).get().getEmail() != null
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(paymentExternalId);
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+
         assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getGatewayPayoutId(), is(payoutId));
+        assertThat(transaction.get().getExternalId(), is(externalId));
+        assertThat(transaction.get().getEmail(), is("test@example.org"));
     }
 
     public void setMessage(byte[] messageContents) {
